# swagger/all.yaml
openapi: 3.0.0
info:
  description: |
    A more complex version of the race cars API.
  version: "1.0.0"
  title: Race Cars Management
  contact:
    email: loi.tra@9thwonder.com
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/loi.tra/my-first-API/1.0.0
  - description: Please run the app locally on port 8080
    url: 'http://localhost:8080/api'
...
paths:
...
  /classes:
    post:
      tags:
        - class
      summary: Add a new class to the system
      operationId: addClass
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassFormData'
        description: form data object that contains fields to update a specific Class
        required: true
      responses:
        '201':
          $ref: '#/components/responses/201Created'
        '400':
          $ref: '#/components/responses/400BadRequest'
    get:
      tags:
        - class
      summary: Finds Classes # with or w/o conditions
      description: Finds all the classes
      # Multiple status values can be provided with comma separated strings
      operationId: findClasses
      parameters:
        - name: pageSize
          $ref: '#/components/parameters/pageSizeParam'
        - name: currentPage
          $ref: '#/components/parameters/currentPageParam'
        # - name: sort
        #   in: query
        #   description: sort the cars returned
        #   schema:
        #     type: string
        #     default: '-createdAt'
        # - name: status
        #   in: query
        #   description: Status values that need to be considered for filter
        #   # required: true
        #   explode: true
        #   schema:
        #     type: array
        #     items:
        #       type: string
        #       enum:
        #         - all
        #         - active
        #         - inactive
        #         - archived
        #       default: all
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Class'
                  count:
                    type: integer
                    format: int32
  '/classes/{id}':
    get:
      tags:
        - class
      summary: Find class by ID
      description: Returns a single class
      operationId: getClassById
      parameters:
        - name: id
          in: path
          description: ID of class to return
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
    patch:
      tags:
        - class
      summary: Updates an existing class by ID
      operationId: updateClass
      parameters:
        - name: id
          in: path
          description: ID of the class that needs to be updated
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassFormData'
        description: form data object that contains fields to update a specific Class
        required: true
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
    delete:
      tags:
        - class
      summary: Deletes a class by ID
      # description: >-
      #   For valid response try integer IDs with positive integer value.\ \
      #   Negative or non-integer values will generate API errors
      operationId: deleteClass
      parameters:
        - name: id
          in: path
          description: Class id to delete
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'

  # Addresses
...
  /drivers:
    post:
      tags:
        - driver
      summary: Add a new driver to the system
      operationId: addDriver
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverFormData'
        description: form data object that contains fields to update a specific driver in the system
      responses:
        '201':
          $ref: '#/components/responses/201Created'
        '400':
          $ref: '#/components/responses/400BadRequest'
    get:
      tags:
        - driver
      summary: Finds Drivers
      description: Finds all the drivers
      operationId: findDrivers
      parameters:
        - name: pageSize
          $ref: '#/components/parameters/pageSizeParam'
        - name: currentPage
          $ref: '#/components/parameters/currentPageParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DriverResponse'
                  count:
                    type: integer
                    format: int32
  '/drivers/{id}':
    get:
      tags:
        - driver
      summary: Find driver by ID
      description: Returns a single driver
      operationId: getDriverById
      parameters:
        - name: id
          in: path
          description: ID of driver to return
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverDetail'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
    patch:
      tags:
        - driver
      summary: Updates an existing driver by ID
      operationId: updateDriver
      parameters:
        - name: id
          in: path
          description: ID of the driver that needs to be updated
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverFormData'
        description: form data object that contains fields to update a specific driver in the system
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
    delete:
      tags:
        - driver
      summary: Deletes a driver by ID
      operationId: deleteDriver
      parameters:
        - name: id
          in: path
          description: Driver id to delete
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
      responses:
        '204':
          $ref: '#/components/responses/204NoContent'
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
  '/drivers/{id}/results':
    get:
      tags:
        - driver
        - race-result
      summary: Get driver's results on races
      description: Get driver's results on races
      operationId: getDriverRaceResults
      parameters:
        - name: id
          in: path
          description: ID of the driver
          required: true
          schema:
            $ref: '#/components/schemas/uuidFormat'
        - name: carId
          in: query
          description: filter the results by a car id
          schema:
            $ref: '#/components/schemas/uuidFormat'
        - name: raceId
          in: query
          description: filter the results by a race id
          schema:
            $ref: '#/components/schemas/uuidFormat'
        - name: pageSize
          $ref: '#/components/parameters/pageSizeParam'
        - name: currentPage
          $ref: '#/components/parameters/currentPageParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RaceResultRespone'
                  count:
                    type: integer
                    format: int32
        '404':
          $ref: '#/components/responses/404NotFound'
        '400':
          $ref: '#/components/responses/400BadRequest'
  
  # Cars
...
components:
  schemas:
    ClassFormData: # Can be referenced as '#/components/schemas/ClassFormData'
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: GTE-Pro
    Class: # Can be referenced as '#/components/schemas/Class'
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/uuidFormat'
        - $ref: '#/components/schemas/ClassFormData'
    AddressFormData: # Can be referenced as '#/components/schemas/AddressFormData'
      type: object
      required:
        - name
        - city
        - state
        - zip
        - country
      properties:
        name:
          type: string
          example: Team A's headquarter
        street:
          type: string
          example: PO BOX 21345
        street2:
          type: string
          example: Street 2
        city:
          type: string
          example: COLLEGE PLACE
        state:
          type: string
          example: WASHINGTON
        zip:
          type: string
          example: "21345"
        country:
          type: string
          example: UNITED STATES
    Address: # Can be referenced as '#/components/schemas/Address'
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/uuidFormat'
        - $ref: '#/components/schemas/AddressFormData'
    TeamFormData: # Can be referenced as '#/components/schemas/TeamFormData'
      type: object
      required:
        - name
        - nationality
      properties:
        name:
          type: string
          example: The Quicksilver
        nationality:
          type: string
          enum: [USA, VietNam]
          example: USA
        businessAddress:
          $ref: '#/components/schemas/uuidFormat'
    Team: # Can be referenced as '#/components/schemas/Team'
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/uuidFormat'
        - $ref: '#/components/schemas/TeamFormData'
    TeamResponse: # Can be referenced as '#/components/schemas/TeamResponse'
      type: object
      required:
        - name
        - nationality
      properties:
        id:
          $ref: '#/components/schemas/uuidFormat'
        name:
          type: string
          example: The Quicksilver
        nationality:
          type: string
          enum: [USA, VietNam]
          example: USA
        businessAddress:
          $ref: '#/components/schemas/Address'
...
    DriverFormData: # Can be referenced as '#/components/schemas/DriverFormData'
      type: object
      required:
        - firstName
        - lastName
      properties:
        firstName:
          type: string
          example: Barry
        lastName:
          type: string
          example: Allen
        nationality:
          type: string
          enum: [USA, VietNam]
          example: USA
        teams:
          type: array
          items:
            $ref: '#/components/schemas/uuidFormat'
        homeAddress:
          $ref: '#/components/schemas/uuidFormat'
        managementAddress:
          $ref: '#/components/schemas/uuidFormat'
    Driver: # Can be referenced as '#/components/schemas/Driver'
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/uuidFormat'
        - $ref: '#/components/schemas/DriverFormData'
    DriverResponse: # Can be referenced as '#/components/schemas/DriverResponse'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuidFormat'
        firstName:
          type: string
          example: Barry
        lastName:
          type: string
          example: Allen
        nationality:
          type: string
          enum: [USA, VietNam]
          example: USA
        team:
          $ref: '#/components/schemas/TeamResponse'
        homeAddress:
          $ref: '#/components/schemas/Address'
        managementAddress:
          $ref: '#/components/schemas/Address'
    DriverDetail: # Can be referenced as '#/components/schemas/DriverDetail'
      allOf:
        - $ref: '#/components/schemas/DriverResponse'
        - type: object
          properties:
            cars:
              $ref: '#/components/schemas/CarResponse'
...
    CarResponse: # Can be referenced as '#/components/schemas/CarResponse'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuidFormat'
        name:
          type: string
          example: A random car's name
        team:
          $ref: '#/components/schemas/Team'
        make:
          type: string
          example: Ferrari
        model:
          type: string
          example: 488 GTE-LM
        class:
          $ref: '#/components/schemas/Class'
...
    RaceFormData: # Can be referenced as '#/components/schemas/RaceFormData'
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Le Mans
...
    Race: # Can be referenced as '#/components/schemas/Race'
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/uuidFormat'
        - $ref: '#/components/schemas/RaceFormData'
...
    RaceResultRespone: # Can be referenced as '#/components/schemas/RaceResultRespone'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuidFormat'
        car:
          $ref: '#/components/schemas/CarResponse'
        carNumber:
          type: integer
          example: 1
        race:
          $ref: '#/components/schemas/Race'
        driver:
          $ref: '#/components/schemas/Driver'
        class:
          $ref: '#/components/schemas/Class'
        startingPosition:
          type: integer
          example: 15
        isFinished:
          type: boolean
          example: true
        finishingPosition:
          type: integer
          example: 3
    uuidFormat: # Can be referenced as '#/components/schemas/uuidFormat'
      type: string
      format: uuid
      minLength: 36
      maxLength: 36
      example: 5f36215d-f5c6-4896-087c-f30f3678f607
  parameters:
    pageSizeParam: # Can be referenced as '#/components/parameters/pageSizeParam'
      name: pageSize
      in: query
      description: aka items per page. Limit the number of resource returned. For valid response try integer IDs with value >= 1 and <= 50.\ \ Other values will generated exceptions
      required: false
      explode: true
      schema:
        type: integer
        format: int32
        default: 10
        minimum: 1
        maximum: 50
      examples:
        min:
          value: 1
          summary: A sample of min value
        max:
          value: 50
          summary: A sample of max value
    currentPageParam: # Can be referenced as '#/components/parameters/currentPageParam'
      name: currentPage
      in: query
      description: offset for pagination
      schema:
        type: integer
        format: int32
        default: 1
        minimum: 1
  responses:
    201Created: # Can be referenced as '#/components/responses/201Created'
      description: Created. The resource was created successfully.
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                $ref: '#/components/schemas/uuidFormat'
    204NoContent: # Can be referenced as '#/components/responses/204NoContent'
      description: No Content
    404NotFound: # Can be referenced as '#/components/responses/404NotFound'
      description: Not found. The resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 404
              error:
                type: string
                example: Not Found
              message:
                type: string
                example: Not Found
    400BadRequest: # Can be referenced as '#/components/responses/400BadRequest'
      description: Bad request. Invalid input supplied.
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 400
              error:
                type: string
                example: Bad Request
              message:
                type: string
                example: '"name" is required'